+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # rollup-pluginutils changelog
 
 ## 2.1.0
-*unreleased*
+*2017-05-07*
 * Add `dataToEsm` helper to create named exports from objects ([#29](https://github.com/rollup/rollup-pluginutils/issues/29))
 * Support literal keys in object patterns ([#27](https://github.com/rollup/rollup-pluginutils/issues/27))
 * Support function declarations without id in `attachScopes` ([#28](https://github.com/rollup/rollup-pluginutils/issues/28))